fix(node/fs): use primordials in _fs_close, _fs_fsync, _fs_fdatasync
Extract close, closeSync, fsync, fsyncSync, fdatasync, and fdatasyncSync
functions into separate files with proper primordials imports.

Part of #24236
